New doc/samples
Here we add excerpts from the rules of National Research University -
Higher School of Economics to the doc/sample folder, as well as a
set of keyphrases associated with it.

This set of texts produces a quite nice output while being analysed
with the corresponding keyphrases (keyphrases graph construction).
It not only makes sense but also differs a bit when turning on
synonym extraction, thus proving the significance of this process.