Return matching scores for all suffixes in AST.score()
The score() method now accepts a new 'return_suffix_scores' parameter (False
by default). When this parameter is set to True, the method returns a pair
(score, suffix_scores), where score is the resulting relevance score of the
given string and suffix_scores is a dict of form {suffix: score}, containing
the scores for each of the suffixes of the input string (they are used while
calculating the integral score bu may be useful on their own as well).