Add support for various relevance measures
The EAST package supported only the AST relevance measure while computing
relevance graphs and keyphrase tables. It was a limit on the range of experiments
facilitated by this software.

Here, we add the RelevanceMeasure class hierarchy that includes two subclasses
so far: the ones for the AST relevance measure and for the cosine similarity.
These classes implement methods set_text_collection() and relevance() that
are used by the main module to compute keyphrase tables/graphs.

There are some changes in the CLI:
* The semantics of the -s option has changed: it is now responsible for
  the relevance measure to use and is no longer a binary option that defines
  whether synonims should be taken into account. Its possible values are now
  "ast" and "cosine"
* The synonimizer options is now -y instead of -s.
* It makes now no sense to use the -d and -a options if -s is set to "cosine"
  (-d and -a are AST-specific).

There are some TODO places in the code that need to be addressed before this
code goes into the next release (support of stemming for languages other than
English, lemmatization, docstring, README, better CLI).