Add support for different languages in cosine similarity
In general, the information about the language in which the texts and
the keyphrases are written is not needed to compute the AST similarity scores.
However, it is used to compute the cosine similarity scores (in case the user
prefers this relevance measure type).

We add the -l parameter to the CLI which enables the manual specification
of the language of the text collection ("english" by default). In particular,
the value of this option controls which Snowball Stemmer is used while
computing the cosine similarity scores.

Bugfixes: the order in which keyphrases were processed (converted to unicode
          etc.) turned out to be wrong; here we fix it.

TODO: consider the possibility of auto language detection.