Add support for single-file text collecitons
If the user provides not a directory but a single .txt file as the input text
collection, this file will be now split by EAST into single lines, and each line
will be considered as a separate text in the collection.

This is particularly useful when one works with short texts collected into
a single file (e.g. tweets).

Also move the numpy import statement in relevance.py to the top (as it seems like
we can add it to the requirements of EAST).